backend/requirements.txt
playwright==1.40.0
google-generativeai>=0.8.3
python-dotenv
Pillow
fastapi
uvicorn
pydantic
apscheduler

